# artifacts/api-server/src/routes/tcc/linear.ts
import { Router } from "express";
import { linear } from "@workspace/integrations-linear";

const router = Router();

router.get("/linear/me", async (_req, res) => {
  try {
    const me = await linear.viewer;
    res.json({
      id: me.id,
      name: me.name,
      email: me.email,
      displayName: me.displayName,
    });
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to fetch user" });
  }
});

router.get("/linear/teams", async (_req, res) => {
  try {
    const teams = await linear.teams();
    const nodes = await teams.nodes;
    res.json(
      nodes.map((t) => ({
        id: t.id,
        name: t.name,
        key: t.key,
        description: t.description ?? null,
      })),
    );
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to fetch teams" });
  }
});

router.get("/linear/projects", async (_req, res) => {
  try {
    const projects = await linear.projects();
    const nodes = await projects.nodes;
    res.json(
      nodes.map((p) => ({
        id: p.id,
        name: p.name,
        description: p.description ?? null,
        state: p.state,
        progress: p.progress,
      })),
    );
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to fetch projects" });
  }
});

router.get("/linear/issues", async (req, res) => {
  try {
    const { teamId, assigneeId, state, first } = req.query as Record<string, string | undefined>;
    const filter: Record<string, any> = {};
    if (teamId) filter.team = { id: { eq: teamId } };
    if (assigneeId) filter.assignee = { id: { eq: assigneeId } };
    if (state) filter.state = { name: { eq: state } };

    const issues = await linear.issues({
      filter: Object.keys(filter).length ? filter : undefined,
      first: first ? parseInt(first, 10) : 50,
      orderBy: "updatedAt" as any,
    });

    const nodes = await Promise.all(
      issues.nodes.map(async (issue) => {
        const state = await issue.state;
        const assignee = await issue.assignee;
        const team = await issue.team;
        return {
          id: issue.id,
          identifier: issue.identifier,
          title: issue.title,
          description: issue.description ?? null,
          priority: issue.priority,
          priorityLabel: issue.priorityLabel,
          url: issue.url,
          state: state ? { id: state.id, name: state.name, color: state.color, type: state.type } : null,
          assignee: assignee ? { id: assignee.id, name: assignee.name, email: assignee.email } : null,
          team: team ? { id: team.id, name: team.name, key: team.key } : null,
          createdAt: issue.createdAt,
          updatedAt: issue.updatedAt,
          dueDate: issue.dueDate ?? null,
        };
      }),
    );

    res.json(nodes);
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to fetch issues" });
  }
});

router.get("/linear/issues/:id", async (req, res) => {
  try {
    const issue = await linear.issue(req.params.id);
    const state = await issue.state;
    const assignee = await issue.assignee;
    const team = await issue.team;
    res.json({
      id: issue.id,
      identifier: issue.identifier,
      title: issue.title,
      description: issue.description ?? null,
      priority: issue.priority,
      priorityLabel: issue.priorityLabel,
      url: issue.url,
      state: state ? { id: state.id, name: state.name, color: state.color, type: state.type } : null,
      assignee: assignee ? { id: assignee.id, name: assignee.name, email: assignee.email } : null,
      team: team ? { id: team.id, name: team.name, key: team.key } : null,
      createdAt: issue.createdAt,
      updatedAt: issue.updatedAt,
      dueDate: issue.dueDate ?? null,
    });
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to fetch issue" });
  }
});

router.post("/linear/issues", async (req, res) => {
  try {
    const { teamId, title, description, priority, assigneeId, dueDate } = req.body as {
      teamId: string;
      title: string;
      description?: string;
      priority?: number;
      assigneeId?: string;
      dueDate?: string;
    };

    if (!teamId || !title) {
      return res.status(400).json({ error: "teamId and title are required" });
    }

    const result = await linear.createIssue({
      teamId,
      title,
      description,
      priority,
      assigneeId,
      dueDate,
    });

    const issue = await result.issue;
    if (!issue) return res.status(500).json({ error: "Issue creation failed" });

    res.status(201).json({ id: issue.id, identifier: issue.identifier, url: issue.url, title: issue.title });
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to create issue" });
  }
});

router.patch("/linear/issues/:id", async (req, res) => {
  try {
    const { title, description, priority, assigneeId, stateId, dueDate } = req.body as {
      title?: string;
      description?: string;
      priority?: number;
      assigneeId?: string;
      stateId?: string;
      dueDate?: string;
    };

    const result = await linear.updateIssue(req.params.id, {
      title,
      description,
      priority,
      assigneeId,
      stateId,
      dueDate,
    });

    const issue = await result.issue;
    if (!issue) return res.status(500).json({ error: "Issue update failed" });

    res.json({ id: issue.id, identifier: issue.identifier, url: issue.url, title: issue.title });
  } catch (err: any) {
    res.status(500).json({ error: err?.message ?? "Failed to update issue" });
  }
});

export default router;