# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Tony's Command Center
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /linear/issues:
    get:
      operationId: getLinearIssues
      tags: [linear]
      summary: List Linear issues
      parameters:
        - in: query
          name: teamId
          schema:
            type: string
          description: Filter by team ID
        - in: query
          name: assigneeId
          schema:
            type: string
          description: Filter by assignee ID
        - in: query
          name: state
          schema:
            type: string
          description: Filter by state name
        - in: query
          name: first
          schema:
            type: integer
          description: Max number of issues to return (default 50)
      responses:
        "200":
          description: Array of issues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LinearIssue"

    post:
      operationId: createLinearIssue
      tags: [linear]
      summary: Create a new Linear issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinearIssueCreateBody"
      responses:
        "201":
          description: Created issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinearIssueRef"
  /linear/issues/{id}:
    get:
      operationId: getLinearIssue
      tags: [linear]
      summary: Get a single Linear issue by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Linear issue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinearIssue"

    patch:
      operationId: updateLinearIssue
      tags: [linear]
      summary: Update a Linear issue
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinearIssueUpdateBody"
      responses:
        "200":
          description: Updated issue reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinearIssueRef"
components:
  schemas:
...
    LinearIssueState:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
        type:
          type: string
    LinearIssueAssignee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
    LinearIssueTeam:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
    LinearIssue:
      type: object
      properties:
        id:
          type: string
        identifier:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: number
        priorityLabel:
          type: string
        url:
          type: string
        state:
          oneOf:
            - $ref: "#/components/schemas/LinearIssueState"
            - type: "null"
        assignee:
          oneOf:
            - $ref: "#/components/schemas/LinearIssueAssignee"
            - type: "null"
        team:
          oneOf:
            - $ref: "#/components/schemas/LinearIssueTeam"
            - type: "null"
        createdAt:
          type: string
        updatedAt:
          type: string
        dueDate:
          type: ["string", "null"]
    LinearIssueRef:
      type: object
      properties:
        id:
          type: string
        identifier:
          type: string
        url:
          type: string
        title:
          type: string
    LinearIssueCreateBody:
      type: object
      required:
        - teamId
        - title
      properties:
        teamId:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
          description: "0=No, 1=Urgent, 2=High, 3=Medium, 4=Low"
        assigneeId:
          type: string
        dueDate:
          type: string
    LinearIssueUpdateBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
        assigneeId:
          type: string
        stateId:
          type: string
        dueDate:
          type: string